@everyone assignments.
By far the most frequently asked questions we get about the permissions system is because DJ is automatically assigned to everyone. Hopefully, the above additions will help reduce the amount of questions regarding the @everyone assignment because, in general, most of these people just want their users to be set to user.