Update proposal for donate.md (#7)
* Update proposal for donate.md

I've been thinking about this page for awhile since I am constantly advertising for Patron in the server. Here's a friendly recommendation for the donation page. The $400 amount number is just an assumption based on the amount of donations coming in on Patreon at the moment.

* Updated with feedback changes

Is this any better?

* Update donate.md #3

Hopefully this is a little more clear and effectively satisfies both of your thoughts and concerns :). Let me know!