Permission Ranks Table (#10)
* Permission Ranks Table

I have noticed people have been frequently confused by the permissions system the way it is laid out as the rank description is under the actual rank. So, I formatted them to be side by side in a table so it may be more easily understood. Included a quick refinement of the notes bullet including the default dj rank, and a quick grammar check at the end.

* Removed Code Frags

Removed the code fragments for now. Should look more normal, then.

* Css Padding Fix Branch9

Fixed the code style padding and put it in 9 to be comparable with the permissions page changes. Will put in a new commit in the permissions markdown to use code blocks in the descriptions of perm ranks.

* Added Code blocks back

There. This should be easier to review and should fix the css issues. Sorry about the confusion. Had to figure out how to commit to the same branch. :D