Streams are supported on YouTube, yes? (#4)
As far as I can tell (and have tested), streams are working. Can we remove this?

Also, just edited a little weird grammar bit. Also, corrected a redundancy. Changed the work 'skips' to 'remove' for the sake of clarity. People assume the queue is a playlist and that, since there is no 'remove' option, it wouldn't remove the list. I think this is primarily an issue for limited English speaking users.